User group minor fixes see #190

+++ main/social/group_members.php
@@ -12,9 +12,7 @@ require_once api_get_path(LIBRARY_PATH).'usermanager.lib.php';
 require_once api_get_path(LIBRARY_PATH).'social.lib.php';
 
 $this_section = SECTION_SOCIAL;
-
 $interbreadcrumb[]= array ('url' =>'home.php','name' => get_lang('Social'));
-
 api_block_anonymous_users();
 
 $group_id	= intval($_GET['id']);
@@ -27,9 +25,8 @@ if (empty($group_id)) {
 	if (empty($group_info)) {
 		api_not_allowed();
 	}
-	//only admin or moderator can do that
 	$user_role = GroupPortalManager::get_user_group_role(api_get_user_id(), $group_id);
-	if (!in_array($user_role, array(GROUP_USER_PERMISSION_ADMIN, GROUP_USER_PERMISSION_MODERATOR))) {
+	if (!in_array($user_role, array(GROUP_USER_PERMISSION_ADMIN, GROUP_USER_PERMISSION_MODERATOR, GROUP_USER_PERMISSION_READER))) {
 		api_not_allowed();		
 	}
 }
@@ -41,15 +38,15 @@ echo '<div class="actions-title">';
 echo get_lang('GroupMembers');
 echo '</div>'; 
 
-// Group information
-$admins		= GroupPortalManager::get_users_by_group($group_id, true,array(GROUP_USER_PERMISSION_ADMIN));
-$show_message = ''; 
+$show_message	= ''; 
+
+//if i'm a moderator
 
 if (isset($_GET['action']) && $_GET['action']=='add') {
 	// we add a user only if is a open group
 	$user_join = intval($_GET['u']);
 	//if i'm a moderator		
-	if (GroupPortalManager::is_group_moderator(api_get_user_id())) {
+	if (GroupPortalManager::is_group_moderator($group_id)) {
 		GroupPortalManager::update_user_role($user_join, $group_id);
 		$show_message = get_lang('UserAdded');
 	}	
@@ -59,7 +56,7 @@ if (isset($_GET['action']) && $_GET['action']=='delete') {
 	// we add a user only if is a open group
 	$user_join = intval($_GET['u']);
 	//if i'm a moderator		
-	if (GroupPortalManager::is_group_moderator(api_get_user_id())) {
+	if (GroupPortalManager::is_group_moderator($group_id)) {
 		GroupPortalManager::delete_user_rel_group($user_join, $group_id); 
 		$show_message = get_lang('UserDeleted');
 	}
@@ -69,7 +66,7 @@ if (isset($_GET['action']) && $_GET['action']=='set_moderator') {
 	// we add a user only if is a open group
 	$user_moderator= intval($_GET['u']);
 	//if i'm the admin		
-	if (GroupPortalManager::is_group_admin(api_get_user_id())) {
+	if (GroupPortalManager::is_group_admin($group_id)) {
 		GroupPortalManager::update_user_role($user_moderator, $group_id, GROUP_USER_PERMISSION_MODERATOR); 
 		$show_message = get_lang('UserChangeToModerator');
 	}
@@ -79,56 +76,57 @@ if (isset($_GET['action']) && $_GET['action']=='delete_moderator') {
 	// we add a user only if is a open group
 	$user_moderator= intval($_GET['u']);
 	//only group admins can do that	
-	if (GroupPortalManager::is_group_admin(api_get_user_id())) {	
+	if (GroupPortalManager::is_group_admin($group_id)) {	
 		GroupPortalManager::update_user_role($user_moderator, $group_id, GROUP_USER_PERMISSION_READER); 
 		$show_message = get_lang('UserChangeToReader');
 	}
 }
 
-
-
 if (! empty($show_message)){
 	Display :: display_normal_message($show_message);
 }
 
-$users	= GroupPortalManager::get_users_by_group($group_id, true, array(GROUP_USER_PERMISSION_ADMIN, GROUP_USER_PERMISSION_READER, GROUP_USER_PERMISSION_MODERATOR));
+$users	= GroupPortalManager::get_users_by_group($group_id, true, array(GROUP_USER_PERMISSION_ADMIN, GROUP_USER_PERMISSION_READER, GROUP_USER_PERMISSION_MODERATOR), 0 , 1000);
 $new_member_list = array();
 
+//Shows left column
+echo GroupPortalManager::show_group_column_information($group_id, api_get_user_id());
 
-	//Shows left column
-	echo GroupPortalManager::show_group_column_information($group_id, api_get_user_id());
-	
-	
-	//-- Show message groups
-	
-	echo '<div id="layout_right" style="margin-left: 282px;">';	
-		
-	// Display form
-	foreach($users as $user) {	 
+//-- Show message groups
+
+echo '<div id="layout_right" style="margin-left: 282px;">';
+
+// Display form
+foreach($users as $user) {		
 		switch ($user['relation_type']) {
 			case  GROUP_USER_PERMISSION_ADMIN:
 				$user['link'] = Display::return_icon('admin_star.png', get_lang('Admin'));
 			break;
 			case  GROUP_USER_PERMISSION_READER:
+				if (in_array($user_role, array(GROUP_USER_PERMISSION_ADMIN, GROUP_USER_PERMISSION_MODERATOR))) {
 				$user['link'] = '<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=delete">'.Display::return_icon('del_user_big.gif', get_lang('DeleteFromGroup')).'</a><br />'.
 								'<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=set_moderator">'.Display::return_icon('admins.gif', get_lang('AddModerator')).'</a>';
+				}
 			break;		
 			case  GROUP_USER_PERMISSION_PENDING_INVITATION:
-				$user['link'] = '<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=add">'.Display::return_icon('pending_invitation.png', get_lang('PendingInvitation')).'</a>';
+				$user['link'] = '<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=add">'.Display::return_icon('pending_invitation.png', get_lang('PendingInvitation')).'</a>';					
 			break;
 			case  GROUP_USER_PERMISSION_MODERATOR:
-				$user['link'] = Display::return_icon('moderator_star.png', get_lang('Moderator')).
-				'<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=delete_moderator">'.Display::return_icon('del_user_big.gif', get_lang('DeleteModerator')).'</a>';
+				$user['link'] = Display::return_icon('moderator_star.png', get_lang('Moderator'));
+				//only group admin can manage moderators 
+				if ($user_role == GROUP_USER_PERMISSION_ADMIN) {
+					$user['link'] .='<a href="group_members.php?id='.$group_id.'&u='.$user['user_id'].'&action=delete_moderator">'.Display::return_icon('del_user_big.gif', get_lang('DeleteModerator')).'</a>';
+				}
 			break;				
 		}
-		$new_member_list[] = $user;
-	}
-	
-	if (count($new_member_list) > 0) {			
-		Display::display_sortable_grid('search_users', array(), $new_member_list, array('hide_navigation'=>true, 'per_page' => 100), $query_vars, false, array(true, false, true,true,false,true,true));		
-	}
-	
-	echo '</div>'; // end layout right
+	$new_member_list[] = $user;
+}
+
+if (count($new_member_list) > 0) {			
+	Display::display_sortable_grid('search_users', array(), $new_member_list, array('hide_navigation'=>true, 'per_page' => 100), $query_vars, false, array(true, false, true,true,false,true,true));		
+}
+
+echo '</div>'; // end layout right
 	
 Display :: display_footer();
 ?>
